Make VNET_ASSERT() available with either VNET_DEBUG or INVARIANTS.
Change the syntax to match KASSERT() to allow more flexible panic
messages rather than having a printf with hardcoded arguments
before panic.

Adjust the few assertions we have to the new format (and enhance
the output) as well as making the CURVNET_RESTORE() condition more
restrictive also checking that curvnet is actually set.

Resort the CURVNET_SET* macros in the non-VNET_DEBUG case to match 
the call order of the VNET_DEBUG case.

[git-p4: depot-paths = "//depot/user/bz/vimage/src/": change = 177255]