Update setting and asserting of curvnet in uipc_socket.c. Document the status quo.
While previously we had more of a mix between caller sets vnet and
function sets vnet, most of the (public) functions now take care
of setting curvnet themselves. This not only helps already virtualized
code like NFS but will also help with virtualizing yet undone protocols.
The exceptions either do not need curvnet to be set or are usually called
from context where curvnet is already set. use an assertion or a comment
there to document the case.

Define VNET_SO_ASSERT(), which is a VNET_ASSERT() with only a socket 
argument to print in case curvnet is NULL to ease debugging as curvnet
should be set from so->so_vnet. Note that CURVNET_SET() has an assertion
itself so duplicating them would never trigger.
Make sure that calls through protosw are all covered by either a
VNET_SO_ASSERT() or CURVNET_SET().  Exceptions to that are basically
(*pru_peeraddr) and (*pru_sockaddr) only needing the current so/pcb
context, as well as (*pru_sopoll), which has a comment.

[git-p4: depot-paths = "//depot/user/bz/vimage/src/": change = 177280]