Introduce a per-subsystem list of all instances. For that define the vimage allocator base information in a standard struct that each subsystem could extend.
Provide one global set of locks to pretect the list.
We will only ever need write locks for starting and
stopping instances.  If this tends to become a problem
the locks can be changed to be per-subsystem as well.

Define a VIMAGE_FOREACH that will walk the list of all
instances of a subsystem. In theory new subsystems should
never need their own macro for that but can just go with
struct vimage *v; and VIMAGE_FOREACH().

[git-p4: depot-paths = "//depot/user/bz/vimage/src/": change = 177827]