Add kvm support to read jail information from kmem or coredumps and allow specifying a different kernel.
Note: this forces lkvm support on all libjail consumers but
even if we would put the additions into libkvm, the wrapper
functions would still lead to unresolveable symbols.  I know
of no options that would be MK_LIBKVM but we might consider
adding that?
We cannot put this into the application without duplicating
(parts of) libjail into jls again.

[git-p4: depot-paths = "//depot/user/bz/vimage/src/": change = 177925]