Implement the ddb logic with the callback function around the "each_symbol_nameval" linker kobj set instance with an approrpiate filter to list all objects of a linker set for a virtualized subsystem.
I will print a header with the subsystem name, and the
beginning and end of the linker set.

It will then print a list of:
object name, object size, the location of the master copy
and the location of the instance currently selected for
debugging by either cur<name> or $db_<name>.

Add a
	show vimage vars <struct vimage_subsys *>
ddb function to the vimage virtualization framework.

Update ddb(4) with the new command and correct some olds.

[git-p4: depot-paths = "//depot/user/bz/vimage/src/": change = 177973]