Fix multiple instances of virtualized variable access without a vnet being set (now that we no longer leak it from kldload). Those assumptions basically never were save.
The current locking primitives do not really work out for
proper synchronysation of instances with more than one stack.
Part of the problem is that some things are global while
locking is virtualized per stack and we cannot use one (random)
to lock access to the globals, especially on unloading, as all
the others might still access things.  Add bandaids and XXXs.

Make the eventhandler there actually do the right thing for
vnets.  It's not the most elegant way as ideally we would only
need to do events for the same vnet but the current design
did not easily allow that in ipfw.

With this ipfw would be loadable again with VIMAGE and in
theory also unloadable (I have VNET_SYSUNINTS disabled so hard
to say).  It's not yet expected to do the right things for
anything but vnet0 - especially wrt to NAT.

Sponsored by: The FreeBSD Foundation
Sponsored by: CK Software GmbH

[git-p4: depot-paths = "//depot/user/bz/vimage/src/": change = 180412]