Split v_sysinit and v_sysunint structs.  We cannot change the v_sysinit struct and calling conventions (easily), as we must be able to compile them out to standard sysinits for bringup.
Change the calling convention for the v_sysunint to pass the
entire v_sysunint struct rather than just arg to be able to
easily extend things in the future (arg is currently only used
by vnet_domain_uninit) passing hints for timeouts or to be run
asynchronously, ..

Add macros (still named sysunint like) to actually dynamically
register the teardown order on bringup.  As we still want
subsystems and order for modules to be hooked into the right
slot in the chain and avoid allocations, the classic but
adjusted SYSUNINTs seems to be more straight forward and
complicate things less.

Keep both for the moment.  Reverting the locking changes in
kern_vimage would be easy but will add kthreads first.

Note that compiling out V_SYSUNINITS to normal onces has been
done half-hearted in the past and made me wonder why we never
saw problems on reboot or halt for GENERIC kernels.  The current
leaning is not to do a proper teardown for single-stack kernels
as it would possibly extend reboot time (a lot - you do not want
to wait for 2 hours or even 2 minutes on TCP during reboot).

Sponsored by: The FreeBSD Foundation
Sponsored by: CK Software GmbH

[git-p4: depot-paths = "//depot/user/bz/vimage/src/": change = 180685]