Make if_vmove() static.  It's file local.
Unify the "ident"s for the VNET_SYS(UN)INITs.

Introduce an extra "shutdown" argument to if_detach_internal()
and callers that in{,6}_ifdetach will not try to cleanup 
upper layer protocol information on network stack shutdown. (*)
By the time we'll get there ULPs are gone already.
This is an initial identifcation of the problem and a workaround
on the lower layers untangling and will need to be reviewed with
dedicated work on those layers.

(*) Changes for inet/inet6 to come in extra submits.

Sponsored by: The FreeBSD Foundation
Sponsored by: CK Software GmbH

[git-p4: depot-paths = "//depot/user/bz/vimage/src/": change = 180828]