The way SCTP currrently gets notified for routing changes by a hard coded function call in rtsock.c troubles in cases when SCTP was "finished" already but we still get route deletes as routes and addresses go from interfaces at a later point and call into SCTP hitting a destroyed lock, etc.
Use the already existent flag to catch those calls and in case
the stack is marked "not intiialized", just return.

Set the "not initialized" flag at the beginning of the sctp
cleanup.

Sponsored by: The FreeBSD Foundation
Sponsored by: CK Software GmbH

Note: given this function call, the one for checksum offloading
and killing the kproc SCTP could be an (un)loadable module, even
on FreeBSD.

[git-p4: depot-paths = "//depot/user/bz/vimage/src/": change = 180833]