Add a flags field to struct vimage_subsystem to allow a subsytem to optionally be marked for asynchronous shutdown.  In that case rather than waiting, start a kproc to do the actual shutdown work. In case starting the kproc fails, fall back to a synchronous shutdown which may block for longer rather than leaking all the resources.
Provide the infrastrucutre within kern_jail to make sure that in   
the VIMAGE case and async shutdown, the struct prison * will not be
freed as that may be needed as authorative source for the subsystem
instance pointer.  A callback from the VIMAGE infrastructure will
do that once the refcount hits zero.  Before that the jail will be
fully unlinked and all other resources be freed.

Mark then virtual network stack (vnet) subsystem for async shutdown,
as, at a later point, we may have to wait there on individual layers.

Sponsored by: The FreeBSD Foundation
Sponsored by: CK Software GmbH

[git-p4: depot-paths = "//depot/user/bz/vimage/src/": change = 180918]