Revert struct vimage_sysuninit back to vimage_sysinit.
Fix the "arg" case for uninit, also for modules by an extra
strcut vimage_sysunint_args (which current has a reserved
field in case we'll ever need to version it for example),
and the original arg.  In the VIMAGE case, pass the pointer
of the master copy to the VIMAGE_SYSUNIT macro and do the
maths for the current instance on the fly.  The virtualization
of args will allow us to pass individual options to each
instance for teardown if needed in the future.
In the non-V case just pass the reference to the struct and
normal SYSUNITs on module unload will just work.

Adjust the ddb function to print the extra arg indirection
and possibly further fields in the future.

Adjust VNET_DOMAIN_SET as the only current consumer.  In case
we compile those out, DOMAIN_SET calls a different set of
functions and there is no "uninit", so there is no need for
further changes.

Sponsored by: The FreeBSD Foundation
Sponsored by: CK Software GmbH

[git-p4: depot-paths = "//depot/user/bz/vimage/src/": change = 180943]