Add (*dom_pr_register) and (*dom_pr_unregister) hooks to the domain definition.  This allows pf_proto_register() and pf_proto_unregister() to automatically call domain specific (de)initialization functions for dynamically registered protocols, eg. from modules and hide the details from them, leaving a single handler rather than domain and protocol specific ones.
Hook up legacy IP and IPv6 ip[6]proto_[un]register() calls to
populate/cleanup the ip[6]_protox[] array registration.

This is bascailly a workaround for a design that should be
changed/improved/maybe redone completely to make things work
for now.

X-Note: This might not be comitted to HEAD but modules like 
carp(4) should "implement" the de-coupled logics for now, also
to allow MFCing the changes, which we cannot do with these.

[git-p4: depot-paths = "//depot/user/bz/vimage/src/": change = 183055]